minor mistake corrected

+++ entrypoint.sh
@@ -43,7 +43,7 @@ if [[ ! -d "${repository}/images" ]]
 then
     mkdir -p "${repository}/images"
     check_success "images folder created" "error: cannot create folder"
-else
+elif [[ -f "${repository}/images/stat.svg" ]]
     rm "${repository}/images/stat.svg" 
     check_success "old images removed" "error: cannot remove images"
 fi
